[FIX] Ajustes finais no código
