+++ nfe/__init__.py
@@ -23,5 +23,4 @@ import l10n_br_account
 import res_company
 import res_partner
 import wizard
-import model
 import report
