+++ nfe/__openerp__.py
@@ -50,7 +50,7 @@ arquivos XML""",
              'account_invoice_view.xml',
              'res_partner_view.xml',
              'res_company_view.xml',
-             "report_print_button_view.xml",
+             'report_print_button_view.xml',
             ],
     'demo': [],
     'test': [],
