+++ nfe/account_invoice.py
@@ -18,16 +18,18 @@
 ###############################################################################
 
 import os
-import datetime
+import pysped
 import logging
+import datetime
 from openerp.osv import orm
+from StringIO import StringIO
 from openerp.tools.translate import _
-from openerp.addons.nfe.sped.nfe.validator.config_check import validate_nfe_configuration, validate_invoice_cancel
-from openerp.addons.nfe.sped.nfe.processing.xml import monta_caminho_nfe
-from openerp.addons.nfe.sped.nfe.processing.xml import send, cancel
-
 from .sped.nfe.nfe_factory import NfeFactory
 from .sped.nfe.validator.xml import XMLValidator
+from openerp.report.pyPdf import PdfFileReader, PdfFileWriter
+from openerp.addons.nfe.sped.nfe.processing.xml import send, cancel
+from openerp.addons.nfe.sped.nfe.processing.xml import monta_caminho_nfe
+from openerp.addons.nfe.sped.nfe.validator.config_check import validate_nfe_configuration, validate_invoice_cancel
 
 _logger = logging.getLogger(__name__)
 
@@ -256,3 +258,83 @@ class AccountInvoice(orm.Model):
                 _logger.error(_(u'Invoice in invalid state to cancel online'),exc_info=True)
                 #TODO
         return
+
+    def str_pdf_Danfes(self, cr, uid, ids, context=None):
+
+        str_pdf = ""
+        paths = []
+
+        for inv in self.browse(cr, uid, ids):
+
+            company = self.pool.get('res.company').browse(
+                cr, uid, inv.company_id.id)
+            nfe_key = inv.nfe_access_key
+            procnfe = pysped.nfe.manual_401.ProcNFe_200()
+
+            try:
+                if inv.state in ['open', 'paid', 'sefaz_cancelled']:
+                    file_xml = os.path.join(monta_caminho_nfe(
+                        company, chave_nfe=nfe_key))
+
+                else:
+                    file_xml = os.path.join(os.path.join(
+                        monta_caminho_nfe(company, chave_nfe=nfe_key), 'tmp/'))
+
+                procnfe.xml = os.path.join(file_xml, nfe_key + '-nfe.xml')
+            except:
+                raise orm.except_orm(('Error !'),
+                                     ('Não é possível gerar a Danfe '
+                                      '- Confirme o Documento'))
+
+            danfe = pysped.nfe.processador_nfe.DANFE()
+            danfe.NFe = procnfe.NFe
+            danfe.protNFe = procnfe.protNFe
+            danfe.caminho = "/tmp/"
+            danfe.gerar_danfe()
+            paths.append(danfe.caminho + danfe.NFe.chave + '.pdf')
+
+        output = PdfFileWriter()
+        s = StringIO()
+
+        # merge dos pdf criados
+        for path in paths:
+            pdf = PdfFileReader(file(path, "rb"))
+
+            for i in range(pdf.getNumPages()):
+                output.addPage(pdf.getPage(i))
+
+            output.write(s)
+
+        str_pdf = s.getvalue()
+        s.close()
+
+        return str_pdf
+
+    def invoice_print(self, cr, uid, ids, context=None):
+
+        for inv in self.browse(cr, uid, ids, context):
+
+            document_serie_id = inv.document_serie_id
+            fiscal_document_id = inv.document_serie_id.fiscal_document_id
+            electronic = inv.document_serie_id.fiscal_document_id.electronic
+
+            if (document_serie_id and fiscal_document_id and not electronic):
+                return super(AccountInvoice, self).invoice_print(
+                    cr, uid, ids, context=context)
+
+            assert len(ids) == 1, 'This option should only be ' \
+                                  'used for a single id at a time.'
+
+            self.write(cr, uid, ids, {'sent': True}, context=context)
+            datas = {
+                'ids': ids,
+                'model': 'account.invoice',
+                'form': self.read(cr, uid, ids[0], context=context)
+            }
+
+            return {
+                'type': 'ir.actions.report.xml',
+                'report_name': 'danfe_account_invoice',
+                'datas': datas,
+                'nodestroy': True
+            }
