--- nfe/model/__init__.py
@@ -1,2 +1,0 @@
-# -*- encoding: utf-8 -*-
-import account_invoice
