+++ nfe/report/report.py
@@ -1,31 +1,26 @@
-# -*- encoding: utf-8 -*-
-##############################################################################
-#
-#    Account Payment Partner module for OpenERP
-#    Copyright (C) 2012 KMEE (http://www.kmee.com.br)
-#    @author Luis Felipe Miléo <mileo@kmee.com.br>
-#
-#    This prog
-# ram is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+    # -*- encoding: utf-8 -*-
+###############################################################################
+#                                                                             #
+# Copyright (C) 2014 Rodolfo Leme Bertozo - KMEE - www.kmee.com.br            #
+#                                                                             #
+#This program is free software: you can redistribute it and/or modify         #
+#it under the terms of the GNU Affero General Public License as published by  #
+#the Free Software Foundation, either version 3 of the License, or            #
+#(at your option) any later version.                                          #
+#                                                                             #
+#This program is distributed in the hope that it will be useful,              #
+#but WITHOUT ANY WARRANTY; without even the implied warranty of               #
+#MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the                #
+#GNU General Public License for more details.                                 #
+#                                                                             #
+#You should have received a copy of the GNU General Public License            #
+#along with this program.  If not, see <http://www.gnu.org/licenses/>.        #
+###############################################################################
 
 from __future__ import with_statement
 from openerp.report.render import render
 from openerp.report.interface import report_int
 from openerp import pooler
-from ..model.account_invoice import account_invoice
 
 
 class external_pdf(render):
@@ -38,13 +33,13 @@ class external_pdf(render):
     def _render(self):
         return self.pdf
 
+
 class report_custom(report_int):
     '''
         Custom report for return danfe
     '''
     def create(self, cr, uid, ids, datas, context={}):
         active_ids = context.get('active_ids')
-        active_model = context.get('active_model')
         pool = pooler.get_pool(cr.dbname)
         list_account_invoice = []
